+++ CHANGELOG.md
@@ -52,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4057](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4057))
 - `opentelemetry-instrumentation-cassandra`: Replace SpanAttributes with semconv constants for DB attributes
   ([#4055](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4055))
+- `opentelemetry-instrumentation-pika`: Replace SpanAttributes with semconv constants for net attributes
+  ([#4068](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4068))
 - `opentelemetry-instrumentation-mysqlclient`: Replace SpanAttributes with semconv constants
   ([#4067](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4067))